feat: Add backoff injection for HTTPAdapter
We define a module functor that allows as to authogenerate a new
module with backoff strategy that handles rate limits, for any module
that implements ExDadata.HTTPAdapter behaviour.

Here we inject in on client initialization.

Note: We don't exactly know what we should treat as rate limit error,
so for now we treat this way all responses that return 403 status
code. We are going to fix this when we add real adapters.